+++ components/molecules/Popup.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Popup from "reactjs-popup";
+
+export default () => (
+  <Popup trigger={<button> Trigger</button>} position="right center">
+    <div>Popup content here !!</div>
+  </Popup>
+);
