+++ lib/index.js
@@ -12,9 +12,11 @@ export const useQuery = initialUrl => {
       console.log("AXIOS", url);
       const res = await axios(url);
       setData(res.data);
+      console.log(res.data);
     };
     fetchUrl();
   }, [url]);
+
   return data;
 };
