+++ pages/ReviewList.js
@@ -18,7 +18,7 @@ function App(p) {
   const data = useQuery(apiUrl);
 
   useEffect(() => {
-    console.log(data);
+    console.log(res.data);
   }, [data]);
   return (
     <div
